change time type

+++ engine/include/enginecore/utils/types/CoreTypesDefines.h
@@ -26,6 +26,6 @@ using itemcount_t = uint32_t;
 using blockstate_t = uint16_t;
 using light_t = uint16_t;
 
-using etime_t = double;
+using etime_t = float;
 
 #endif // !TYPES_CORETYPESDEFINE_H_
